feat (ads-service): Implement advertisement controller methods Configure request authorization checks Add global exception handler
fix: Adjust image URL response according to requirements

+++ backend/ads-service/src/main/java/com/github/spjavaind300/service/imp/AdServiceImp.java
@@ -1,11 +1,14 @@
 package com.github.spjavaind300.service.imp;
 
+import com.github.spjavaind300.exception.AccessDeniedException;
 import com.github.spjavaind300.exception.NotFoundException;
 import com.github.spjavaind300.model.dto.AdExtraInfoDto;
 import com.github.spjavaind300.model.dto.AdRequestDto;
 import com.github.spjavaind300.model.dto.AdResponseDto;
 import com.github.spjavaind300.model.dto.ImageDto;
 import com.github.spjavaind300.model.dto.ListAdsDto;
+import com.github.spjavaind300.model.dto.Role;
+import com.github.spjavaind300.model.dto.UserContext;
 import com.github.spjavaind300.model.dto.UserDto;
 import com.github.spjavaind300.model.entity.Ad;
 import com.github.spjavaind300.model.mapper.AdMapper;
@@ -19,8 +22,8 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.time.Duration;
 import java.util.List;
+
 
 @Service
 @RequiredArgsConstructor
@@ -39,10 +42,7 @@ public class AdServiceImp implements AdService {
         return ListAdsDto.builder()
                 .count(ads.size())
                 .items(ads.stream()
-                        .map(ad -> adMapper.toAdResponseDto(
-                                ad,
-                                imageStorageService.getPreSignedUrl(ad.getImageKey(), Duration.ofMinutes(15)))
-                        )
+                        .map(adMapper::toAdResponseDto)
                         .toList())
                 .build();
     }
@@ -53,10 +53,7 @@ public class AdServiceImp implements AdService {
         return ListAdsDto.builder()
                 .count(ads.size())
                 .items(ads.stream()
-                        .map(ad -> adMapper.toAdResponseDto(
-                                ad,
-                                imageStorageService.getPreSignedUrl(ad.getImageKey(), Duration.ofMinutes(15)))
-                        )
+                        .map(adMapper::toAdResponseDto)
                         .toList())
                 .build();
     }
@@ -79,24 +76,27 @@ public class AdServiceImp implements AdService {
 
     @Transactional
     @Override
-    public void deleteAd(int id) {
+    public void deleteAd(int id, UserContext userContext) {
         Ad ad = adRepository.findById(id).orElseThrow(() -> new NotFoundException(id));
+        checkUserAccess(ad.getUserId(), userContext);
         imageStorageService.deleteFile(ad.getImageKey());
         adRepository.delete(ad);
     }
 
     @Transactional
     @Override
-    public AdResponseDto updateAd(int id, AdRequestDto adRequestDto) {
+    public AdResponseDto updateAd(int id, AdRequestDto adRequestDto, UserContext userContext) {
         Ad ad = adRepository.findById(id).orElseThrow(() -> new NotFoundException(id));
+        checkUserAccess(ad.getUserId(), userContext);
         Ad updateAd = adMapper.updateAd(ad, adRequestDto);
         return saveAd(updateAd);
     }
 
     @Transactional
     @Override
-    public byte[] updateImage(int id, MultipartFile image) {
+    public byte[] updateImage(int id, MultipartFile image, UserContext userContext) {
         Ad ad = adRepository.findById(id).orElseThrow(() -> new NotFoundException(id));
+        checkUserAccess(ad.getUserId(), userContext);
         ad.setImageKey(imageStorageService.uploadFile(image).url());
         ad.setOriginalFilename(image.getOriginalFilename());
         adRepository.save(ad);
@@ -106,8 +106,13 @@ public class AdServiceImp implements AdService {
     private AdResponseDto saveAd(Ad ad) {
 
         return adMapper.toAdResponseDto(
-                adRepository.save(ad),
-                imageStorageService.getPreSignedUrl(ad.getImageKey(), Duration.ofMinutes(15))
+                adRepository.save(ad)
         );
+    }
+
+    private void checkUserAccess(long userId, UserContext userContext) {
+        if (userContext.role() == Role.USER && userContext.userId() != userId) {
+            throw new AccessDeniedException();
+        }
     }
 }
